import-strava-api/pyproject.toml
[project]
name = "import-strava-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
